Fixed #35. Make compatible with call/cc and CPS.
